refactor: 项目结构调整
