--- user-common/src/main/java/com/dqzhou/common/rpc/UserService.java
@@ -1,10 +1,0 @@
-package com.dqzhou.common.rpc;
-
-import com.dqzhou.common.response.UserInfo;
-
-public interface UserService {
-
-    void checkPassword(String username, String password);
-
-    UserInfo getUserById(int uid);
-}
