+++ user-service/pom.xml
@@ -4,7 +4,7 @@
     <parent>
         <artifactId>ucenter</artifactId>
         <groupId>com.dqzhou</groupId>
-        <version>${project.version}</version>
+        <version>1.0.1.RELEASE</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.dqzhou</groupId>
@@ -21,6 +21,11 @@
         <dependency>
             <groupId>com.dqzhou</groupId>
             <artifactId>user-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.dqzhou</groupId>
+            <artifactId>user-service-facade</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
@@ -79,10 +84,29 @@
     </dependencies>
 
     <build>
+        <finalName>user-center-service</finalName>
         <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.7.0</version>
+                <configuration>
+                    <source>1.8</source>
+                    <target>1.8</target>
+                </configuration>
+            </plugin>
             <plugin>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
+                <configuration>
+                </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-release-plugin</artifactId>
+                <version>2.5.3</version>
+                <configuration>
+                </configuration>
             </plugin>
         </plugins>
     </build>
